requirements.txt
telethon
python-dotenv
fastapi
uvicorn
jinja2
python-multipart
